+++ src/core/types/station.ts
@@ -1,0 +1,8 @@
+import { Coordinates } from './graph';
+
+export interface NearbyStationInfo {
+  id: string;
+  name: string;
+  distance: number;
+  coordinates: Coordinates;
+}
