+++ src/types/error.ts
@@ -1,0 +1,9 @@
+// Define shared error codes - these should match client-side codes
+export const ErrorCodes = {
+  MISSING_PARAMETERS: 'MISSING_PARAMETERS',
+  NO_START_STATION: 'NO_START_STATION',
+  NO_END_STATION: 'NO_END_STATION',
+  NO_ROUTES_FOUND: 'NO_ROUTES_FOUND',
+  SERVER_ERROR: 'SERVER_ERROR',
+  NETWORK_ERROR: 'NETWORK_ERROR', // Client-side only
+};
